references and associators enumeration commands cause an inverse request
In wsmancli-2.2.3, we notice that the references and associators
enumeration commands cause an inverse request.  That is, a
"references" action command adds a "wsmb:AssociatedInstances" in the
enumeration filter.  And an "associators" action command adds a
"wsmb:AssociationInstances" in the enumeration filter.

An "associators" command DSP0200 section 5.3.2.14 enumerates
"associated" classes. This statement is also defined by DSP0227
section 17.1 where it specifies in section 17.1.7 the use of
"wsmb:AssociatedInstances" element in the filter.

A "references" command DSP0200 section 5.3.2.16 enumerates
"association" classes. This statement is also defined by DSP0227
section 17.1 where it specifies in section 17.1.9 the use of
"wsmb:AssociationInstances" element in the filter.