wsmancli allow debugging of invalid response XML
sing wsmancli, when the response XML is invalid (such as due to
invalid characters) an error is displayed with only a partial view of
the XML and no other information that could tell us where exactly the
problem is.

In debug mode (using -d option), it would be useful if the tool could
dump the raw XML for further analysis before it could be interpreted
by libxml functions.

Patch by Chris Poblete